iterating on immutablity API - factoring out API to immutable_ir.py
